bau: Delete WalletAuthoriseService
Move the wallet authorisation into AuthorisationService.
