+++ src/main/java/uk/gov/pay/connector/wallets/googlepay/GooglePayService.java
@@ -1,7 +1,7 @@
 package uk.gov.pay.connector.wallets.googlepay;
 
+import uk.gov.pay.connector.paymentprocessor.service.AuthorisationService;
 import uk.gov.pay.connector.wallets.WalletAuthorisationRequest;
-import uk.gov.pay.connector.wallets.WalletAuthoriseService;
 import uk.gov.pay.connector.wallets.WalletService;
 import uk.gov.pay.connector.wallets.WalletType;
 import uk.gov.pay.connector.wallets.googlepay.api.GooglePayAuthRequest;
@@ -12,7 +12,7 @@ import javax.inject.Inject;
 public class GooglePayService extends WalletService {
     
     @Inject
-    public GooglePayService(WalletAuthoriseService authoriseService) {
+    public GooglePayService(AuthorisationService authoriseService) {
         super(authoriseService, WalletType.GOOGLE_PAY);
     }
